[*] Changed middlewares and tests
