+++ bin/run.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 npm i --quiet
-NODE_END=test npm test
+NODE_ENV=test npm test
