+++ src/auth.service.js
@@ -1,26 +1,26 @@
 const jwt = require('jsonwebtoken');
+const _ = require('lodash');
 const config = require('config');
 
 const { logger } = global;
-
-const jwtOptions = {
-  audience: config.audience,
-  issuer: config.issuer,
-};
 
 exports.createAuthToken = ({ userId }) => {
   const payload = {
     _id: userId,
   };
 
-  return jwt.sign(payload, config.jwtSecret, jwtOptions);
+  return jwt.sign(
+    payload,
+    config.jwt.secret,
+    _.pick(config.jwt, ['audience', 'issuer']),
+  );
 };
 
 exports.decodeToken = (token) => {
   let res;
 
   try {
-    res = jwt.verify(token, config.jwtSecret);
+    res = jwt.verify(token, config.jwt.secret);
   } catch (err) {
     logger.warn('Invalid json web token', err);
   }
