+++ src/config/environment/development.js
@@ -2,6 +2,8 @@ module.exports = {
   mongo: {
     connection: 'mongodb://mongo:27017/api-development',
   },
-  jwtSecret: 'the_secret',
+  jwt: {
+    secret: 'the_secret',
+  },
   authCookieName: 'auth',
 };
