+++ src/config/environment/index.js
@@ -10,9 +10,11 @@ let base = {
   isDev: env === 'development',
   isTest: env === 'test',
   webUrl: 'http://localhost:3002',
-  jwtSecret: 'jwt_secret',
-  audience: 'api',
-  issuer: 'api',
+  jwt: {
+    secret: 'jwt_secret',
+    audience: 'api',
+    issuer: 'api',
+  },
 };
 
 const envConfig = require(`./${env}.js`); // eslint-disable-line
