+++ src/config/routes/authenticated.js
@@ -1,5 +1,5 @@
 const mount = require('koa-mount');
-const userResource = require('resources/user/authenticated');
+const userResource = require('resources/user');
 
 module.exports = (app) => {
   app.use(mount('/users', userResource));
