+++ src/config/routes/index.js
@@ -2,21 +2,15 @@ const jwt = require('koa-jwt');
 
 const config = require('config');
 
-const { urlToken, user } = require('infrastructure/middlewares');
+const { urlToken, user } = require('./middlewares');
 const publicRoutes = require('./public');
 const authenticatedRoutes = require('./authenticated');
-
-const jwtOptions = {
-  secret: config.jwtSecret,
-  audience: config.audience,
-  issuer: config.issuer,
-};
 
 const defineRoutes = (app) => {
   publicRoutes(app);
 
   app.use(urlToken);
-  app.use(jwt(jwtOptions));
+  app.use(jwt(config.jwt));
   app.use(user);
 
   authenticatedRoutes(app);
