--- src/tests/app.spec.js
@@ -1,9 +1,0 @@
-const chai = require('chai');
-
-const server = require('../app');
-const request = require('supertest').agent(server.listen());
-
-chai.should();
-
-require('../resources/account/account.test')(request);
-require('../resources/user/user.test')(request);
